@@ -1,4 +1,4 @@
-class CreateStates < ActiveRecord::Migration[5.1]
+class CreateStates < ActiveRecord::Migration[7.0]
   def change
     create_table :states do |t|
       t.string :name
