@@ -1,4 +1,4 @@
-class CreateCounties < ActiveRecord::Migration[5.1]
+class CreateCounties < ActiveRecord::Migration[7.0]
   def change
     create_table :counties do |t|
       t.string :name
