@@ -1,4 +1,4 @@
-class CreateParties < ActiveRecord::Migration[5.1]
+class CreateParties < ActiveRecord::Migration[7.0]
   def change
     create_table :parties do |t|
       t.string :name
