@@ -1,4 +1,4 @@
-class CreateCandidates < ActiveRecord::Migration[5.1]
+class CreateCandidates < ActiveRecord::Migration[7.0]
   def change
     create_table :candidates do |t|
       t.string :name
