@@ -1,4 +1,4 @@
-class AddPartyToCandidate < ActiveRecord::Migration[5.1]
+class AddPartyToCandidate < ActiveRecord::Migration[7.0]
   def change
     add_reference :candidates, :party, foreign_key: true
   end
