@@ -1,7 +1,7 @@
-require 'rails_helper'
+require "rails_helper"
 
-feature 'When a user visits /states/:id' do
-  scenario 'they see a summary of votes within that state by candidate' do
+RSpec.describe "When a user visits /states/:id", type: :feature do
+  scenario "they see a summary of votes within that state by candidate" do
     state = create(:state)
     candidate1, candidate2 = create_list(:candidate, 2)
     county1, county2 = create_list(:county, 2, state: state)
@@ -14,7 +14,7 @@ feature 'When a user visits /states/:id' do
 
     visit state_path(state)
 
-    within('.summary') do
+    within(".summary") do
       expect(page).to have_content("#{candidate1.name}: #{candidate1_votes}")
       expect(page).to have_content("#{candidate2.name}: #{candidate2_votes}")
     end
