@@ -1,7 +1,7 @@
-require 'rails_helper'
+require "rails_helper"
 
 describe Party do
-  describe 'Validations' do
+  describe "Validations" do
     it { should validate_presence_of(:name) }
     it { should validate_uniqueness_of(:name) }
   end
